refactor: 데이터 초기화
