+++ src/test/java/com/soolsul/soolsulserver/data/DataLoaderBootstrap.java
@@ -1,10 +1,8 @@
 package com.soolsul.soolsulserver.data;
 
-import com.soolsul.soolsulserver.user.auth.business.RoleHierarchyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationListener;
 import org.springframework.context.event.ContextRefreshedEvent;
-import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -12,13 +10,9 @@ import org.springframework.stereotype.Component;
 public class DataLoaderBootstrap implements ApplicationListener<ContextRefreshedEvent> {
 
     private final DataLoader dataLoader;
-    private final RoleHierarchyService roleHierarchyService;
-    private final RoleHierarchyImpl roleHierarchy;
 
     @Override
     public void onApplicationEvent(ContextRefreshedEvent event) {
         dataLoader.loadData();
-        String allHierarchy = roleHierarchyService.AllHierarchyToString();
-        roleHierarchy.setHierarchy(allHierarchy);
     }
 }
